Create wrapper component Card
