+++ src/components/QuestionCard.tsx
@@ -1,6 +1,7 @@
-import { View, Text, StyleSheet } from "react-native";
+import { View } from "react-native";
 import AnswerOption from "./AnswerOption";
 import { Question } from "../types";
+import Card from "./Card";
 
 type QuestionCardProps = {
   question: Question;
@@ -14,8 +15,7 @@ export default function QuestionCard(props: QuestionCardProps) {
   };
 
   return (
-    <View style={styles.questionCard}>
-      <Text style={styles.question}>{props.question.title}</Text>
+    <Card title={props.question.title}>
       <View style={{ gap: 20 }}>
         {props.question.options.map((option, index) => (
           <AnswerOption
@@ -26,21 +26,6 @@ export default function QuestionCard(props: QuestionCardProps) {
           />
         ))}
       </View>
-    </View>
+    </Card>
   );
 }
-
-const styles = StyleSheet.create({
-  questionCard: {
-    backgroundColor: "#FFFFFF",
-    padding: 20,
-    paddingVertical: 40,
-    borderRadius: 20,
-    gap: 20,
-    boxShadow: "0px 0px 10px 0px rgba(0, 0, 0, 0.1)",
-  },
-  question: {
-    fontSize: 18,
-    fontWeight: 500,
-  },
-});
